refactor: nest GraphQL operations under a downloadHelper namespace container
Move the 2 root Query (downloadHelperInfo/Files) + 2 root Mutation (downloadHelperTrigger/
DeleteFile) flat fields under one hierarchical namespace container, per the Jahia GraphQL
convention.

- Query:    downloadHelperInfo -> downloadHelper { info }, ...Files -> { files }
- Mutation: downloadHelperTrigger -> downloadHelper { trigger }, ...DeleteFile -> { deleteFile }

New holder types DownloadHelperQuery/Mutation carry the operations as instance
@GraphQLField methods; the *Extension classes keep a single downloadHelper accessor field.
@GraphQLRequiresPermission stays on each leaf. Frontend gql.js + jsx + the jest test
(DownloadHelper.test.jsx mock shapes) and Cypress fixtures/specs updated in lockstep.
