+++ tests/cypress/fixtures/graphql/mutation/triggerDownload.graphql
@@ -6,12 +6,14 @@ mutation TriggerDownload(
     $password: String
     $email: String
 ) {
-    downloadHelperTrigger(
-        protocol: $protocol
-        url: $url
-        filename: $filename
-        login: $login
-        password: $password
-        email: $email
-    )
+    downloadHelper {
+        trigger(
+            protocol: $protocol
+            url: $url
+            filename: $filename
+            login: $login
+            password: $password
+            email: $email
+        )
+    }
 }
